urban_nav/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
torch==2.3.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
requests==2.31.0
geopy==2.4.1
python-dotenv==1.0.1
schedule==1.2.1
pydantic==2.7.1
httpx==0.27.0
